requirements.txt
APScheduler==3.8.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
visitor==0.1.3
-e git+ssh://git@github.com/pavelkrizek/refactor_watering.git@b92a7ad358c46bb798126b9741cab11be50bc849#egg=water_garden
Werkzeug==2.0.2
wincertstore==0.2
WTForms==3.0.0
